requirements.txt
pymupdf
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
tqdm==4.66.4
